refactored creation of default scene data into provider classes
