+++ UltraStar Play/Assets/Common/Scene/IDefaultSceneDataProvider.cs
@@ -1,0 +1,4 @@
+public interface IDefaultSceneDataProvider
+{
+    SceneData GetDefaultSceneData();
+}
