+++ UltraStar Play/Assets/Common/Scene/IDefaultSceneDataProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ef585037f670164e958ded2a26bdc2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
