+++ UltraStar Play/Assets/Common/Scene/SceneDataException.cs
@@ -1,0 +1,8 @@
+using System;
+
+public class SceneDataException : Exception
+{
+    public SceneDataException(string message) : base(message)
+    {
+    }
+}
