+++ UltraStar Play/Assets/Common/Scene/SceneDataException.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73b5b160440833d49b0b9595c516d9d6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
