+++ UltraStar Play/Assets/Scenes/Sing/DefaultSingSceneDataProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04e8e8353cf7d75419d03321d0098a54
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
