+++ UltraStar Play/Assets/Scenes/Sing/SingScene.unity
@@ -323,7 +323,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &234276797
 MonoBehaviour:
@@ -337,6 +337,52 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 35ef2734292163c45a563380ea54c792, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1 &242235908
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 242235910}
+  - component: {fileID: 242235909}
+  m_Layer: 0
+  m_Name: DefaultSingSceneDataProvider
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &242235909
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 242235908}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 04e8e8353cf7d75419d03321d0098a54, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  defaultSongName: Hakuna matata
+  defaultSongNamePasteBin: "Summer Wine\r\nEmptyTitle\r\nMe And My Broken Heart\r\nSomewhere
+    Over The Rainbow (Wonderful World)\r\nMarry You\r\nHakuna matata"
+--- !u!4 &242235910
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 242235908}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 512.0799, y: 274.07208, z: -458.25}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &242843841
 GameObject:
   m_ObjectHideFlags: 0
@@ -610,11 +656,11 @@ MonoBehaviour:
   anchorMin: {x: 0, y: 0}
   anchorMax: {x: 0, y: 0}
   pivot: {x: 0.5, y: 0.5}
-  localPosition: {x: 492.5, y: 274, z: 0}
-  anchoredPosition: {x: 492.5, y: 274}
-  position: {x: 492.5, y: 274, z: 0}
-  sizeDelta: {x: 799.99994, y: 445.0761}
-  size: {x: 799.99994, y: 445.0761}
+  localPosition: {x: 493.5, y: 274, z: 0}
+  anchoredPosition: {x: 493.5, y: 274}
+  position: {x: 493.5, y: 274, z: 0}
+  sizeDelta: {x: 800, y: 444.17426}
+  size: {x: 800, y: 444.17426}
 --- !u!1 &475691986
 GameObject:
   m_ObjectHideFlags: 0
@@ -889,7 +935,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 7
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!224 &746576793 stripped
 RectTransform:
@@ -1060,7 +1106,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &908260462
 GameObject:
@@ -1105,7 +1151,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &992840886
 GameObject:
@@ -1785,7 +1831,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1343510743
 GameObject:
@@ -1897,18 +1943,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0c11810cb9527c047beda7fd4b488be2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  defaultSongName: Hakuna matata
-  defaultSongNamePasteBin: 'Summer Wine
-
-    EmptyTitle
-
-    Me And My Broken Heart
-
-    Somewhere Over The Rainbow (Wonderful World)
-
-    Marry You
-
-    Hakuna matata'
   pauseOverlay: {fileID: 1219814411}
   playerControllerPrefab: {fileID: 2965445179344377450, guid: 614fa56bea7fef5428e98f6522cee8eb,
     type: 3}
