+++ UltraStar Play/Assets/Scenes/SingingResults/DefaultSingingResultsSceneDataProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71feee30a721fe44abd1b0b42c8ea1d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
