+++ UltraStar Play/Assets/Scenes/SingingResults/SingingResultsSceneController.cs
@@ -24,7 +24,7 @@ public class SingingResultsSceneController : MonoBehaviour
 
     void Start()
     {
-        sceneData = SceneNavigator.Instance.GetSceneData<SingingResultsSceneData>(CreateDefaultSceneData());
+        sceneData = SceneNavigator.Instance.GetSceneDataOrThrow<SingingResultsSceneData>();
         SelectLayout();
         FillLayout();
     }
@@ -72,23 +72,6 @@ public class SingingResultsSceneController : MonoBehaviour
             return twoPlayerLayout;
         }
         return onePlayerLayout;
-    }
-
-    private SingingResultsSceneData CreateDefaultSceneData()
-    {
-        SingingResultsSceneData data = new SingingResultsSceneData();
-        SongMetaManager.Instance.WaitUntilSongScanFinished();
-        data.SongMeta = SongMetaManager.Instance.GetFirstSongMeta();
-
-        SingingResultsSceneData.PlayerScoreData playerScoreData = new SingingResultsSceneData.PlayerScoreData();
-        playerScoreData.TotalScore = 6500;
-        playerScoreData.NormalNotesScore = 4000;
-        playerScoreData.GoldenNotesScore = 2000;
-        playerScoreData.PerfectSentenceBonusScore = 500;
-
-        PlayerProfile playerProfile = SettingsManager.Instance.Settings.PlayerProfiles[0];
-        data.AddPlayerScores(playerProfile, playerScoreData);
-        return data;
     }
 
     public void FinishScene()
