+++ UltraStar Play/Assets/Scenes/SongEditor/DefaultSongEditorSceneDataProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bfac381e8120d2d41b8da847760356a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
